requirements.txt
python-dotenv
pandas
google-generativeai
chromadb
faiss-python